+++ tests/Catalog/SearchProductAttributeTest.php
@@ -9,6 +9,7 @@ namespace Wizaplace\SDK\Tests\Catalog;
 
 use PHPUnit\Framework\TestCase;
 use Wizaplace\SDK\Catalog\AttributeType;
+use Wizaplace\SDK\Catalog\ProductAttributeValue;
 use Wizaplace\SDK\Catalog\SearchProductAttribute;
 
 final class SearchProductAttributeTest extends TestCase
@@ -43,8 +44,8 @@ JSON;
         $this->assertSame('', $attribute->getSlug()); // @FIXME @TODO : the API does not give us this data. It should be added to the API, or removed from the SDK
         $this->assertTrue(AttributeType::CHECKBOX_MULTIPLE()->equals($attribute->getType()));
         $this->assertSame('Couleur test', $attribute->getName());
-        $this->assertSame([
-            [
+        $this->assertEquals([
+            new ProductAttributeValue([
                 'id' => 64,
                 'attributeId' => 52,
                 'name' => 'blue',
@@ -52,7 +53,7 @@ JSON;
                 'image' => [
                     'id' => 708,
                 ],
-            ],
+            ]),
         ], $attribute->getValues());
     }
 }