+++ tests/Catalog/CatalogServiceTest.php
@@ -12,6 +12,7 @@ use Wizaplace\SDK\Catalog\Attribute;
 use Wizaplace\SDK\Catalog\AttributeType;
 use Wizaplace\SDK\Catalog\AttributeVariant;
 use Wizaplace\SDK\Catalog\CatalogService;
+use Wizaplace\SDK\Catalog\Condition;
 use Wizaplace\SDK\Catalog\Declination;
 use Wizaplace\SDK\Catalog\Facet;
 use Wizaplace\SDK\Catalog\Option;
@@ -254,7 +255,8 @@ final class CatalogServiceTest extends ApiTestCase
         $this->assertGreaterThan(1400000000, $product->getCreatedAt()->getTimestamp());
         $this->assertNull($product->getAverageRating());
         $this->assertSame('Product with shippings', $product->getName());
-        $this->assertSame(['N'], $product->getCondition());
+        $this->assertContainsOnly(Condition::class, $product->getConditions());
+        $this->assertEquals([Condition::BRAND_NEW()], $product->getConditions());
         $this->assertSame(1, $product->getDeclinationCount());
         $this->assertSame('', $product->getSubtitle());
         $this->assertSame("La nouvelle génération de notre tablette Fire phare - désormais plus fine, plus légère, dotée d'une plus longue autonomie et d'un écran amélioré.", $product->getShortDescription());