+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 ### BREAKING CHANGES
 
 - `\Wizaplace\SDK\Order\OrderReturn::getStatus` now returns an `\Wizaplace\SDK\Order\OrderReturnStatus` instead of a string
+- `\Wizaplace\SDK\Basket\PaymentInformation::getRedirectUrl` now returns a nullable `\Psr\Http\Message\UriInterface` instead of a `string`
+- `\Wizaplace\SDK\Company\Company::getUrl` now returns a nullable `\Psr\Http\Message\UriInterface` instead of a `string`
+- `\Wizaplace\SDK\Catalog\ProductVideo::getThumbnailUrl` now returns an `\Psr\Http\Message\UriInterface` instead of a `string`
+- `\Wizaplace\SDK\Catalog\ProductVideo::getVideoUrl` now returns an `\Psr\Http\Message\UriInterface` instead of a `string`
+- `\Wizaplace\SDK\Image\ImageService::getImageLink` now returns an `\Psr\Http\Message\UriInterface` instead of a `string`
 
 ### New features