+++ CHANGELOG.md
@@ -23,7 +23,7 @@
 - `\Wizaplace\SDK\Catalog\Facet` has moved to `\Wizaplace\SDK\Catalog\Facet\Facet` and is now an abstract class with two subclasses: `\Wizaplace\SDK\Catalog\Facet\ListFacet` and `\Wizaplace\SDK\Catalog\Facet\NumericFacet`. A few methods have changed in the process
 - `\Wizaplace\SDK\Catalog\DeclinationSummary::getProductId` now returns a `string` instead of an `int`
 - All declinations' IDs were changed from `string` to `\Wizaplace\SDK\Catalog\DeclinationId`
-- Added `\Wizaplace\SDK\Catalog\ProductSummary::getSingleDeclinationId`
+- Added `\Wizaplace\SDK\Catalog\ProductSummary::getMainDeclinationId`
 
 ### New features